fixup! fixup! fixup! fixup! feat: remove dependency on Workspace API
