+++ packages/dashboard-frontend/src/store/FactoryResolver/__tests__/index.spec.ts
@@ -117,9 +117,6 @@ describe('FactoryResolver store', () => {
       getFactoryResolverSpy.mockResolvedValueOnce(resolver);
 
       const store = new FakeStoreBuilder()
-        .withWorkspacesSettings({
-          'che.workspace.storage.preferred_type': 'ephemeral',
-        })
         .build() as MockStoreEnhanced<
         AppState,
         ThunkDispatch<AppState, undefined, factoryResolverStore.KnownAction>
